titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/shared/tep/data-access-direct-tep/src/lib/+state/direct-tep.reducer.ts
import { createFeatureSelector } from '@ngrx/store';
import { TepActionTypes, TepActions } from './direct-tep.actions';
import { TepState } from './direct-tep.state';

export const tepFeatureKey = 'TEP';

export const selectTepState = createFeatureSelector<TepState>(tepFeatureKey);

export const initialState: TepState = {
  errors: null,
  isLoading: false,
  isOpenTaskLoading: false,
  selectedRsoName: null,
  isLoadingPriceUpdate: false,
  selectedCmItem: null,
  createOpenTepTransactionResponse: null,
  updateOpenTepTransactionResponse: null,
  tepItemConfiguratonResponse: null,
  tepCashMemoResponseItemList: null,
  tepPriceDetailsResponse: null,
  addTepItemResponse: null,
  updateTepItemResponse: null,
  confirmTepItemResponse: null,
  deleteTepItemResponse: null,
  rsoList: [],
  cancelResponse: null,
  cancelTEPResponse: null,
  UpdatePriceDetailsResponse: null,
  remarks: '',
  totalQty: 0,
  totalGrossWt: 0,
  selectedData: null,
  refundCashLimit: null,
  totalExchangeAmt: 0,
  selectedPaymentMethod: null,
  selectedTepType: null,
  scannedTepItemCode: null,
  viewTepTransactionResponse: null,
  viewTepItemResponse: null,
  deleteTepTransactionResponse: null,
  tepItemCutPieceDetailsResponse: null,
  cutPieceTotalQty: 0,
  cutPieceTotalValue: 0,
  cmListItemTepConfigurationResponse: null,
  isRefundFormValid: false,
  isRequestRaisingScenario: false,
  goldPlusLocations: [],
  uploadFileResponse: false,
  downloadIdProofFileUrl: null,
  downloadCancelledChequeFileUrl: null,
  downloadApprovalMailFileUrl: null,
  fvtReasons: [],
  updateTepTransactionPriceDetailsResponse: null,
  holdTransactionMetalRates: null,

  createOpenCutPieceTepTransactionResponse: null,
  patchCutPieceTepTransactionResponse: null,
  addCutPieceTepItemResponse: null,
  patchCutPieceTepItemResponse: null,
  confirmCutPieceTepItemResponse: null,
  availableDiscountsList: null,
  studdedProductGroupCodes: [],
  isExceptionScenario: false
};

export function TepReducer(
  state: TepState = initialState,
  action: TepActions
): TepState {
  switch (action.type) {
    case TepActionTypes.CREATE_OPEN_TEP_TRANSACTION:
      return {
        ...state,
        isLoading: true,
        isOpenTaskLoading: true,
        errors: null
      };
    case TepActionTypes.UPDATE_OPEN_TEP_TRANSACTION:
    case TepActionTypes.GET_TEP_ITEM_CONFIGURATION:
    case TepActionTypes.GET_TEP_ITEM_EXCEPTION_CONFIGURATION:

    case TepActionTypes.GET_TEP_CASH_MEMO_ITEM_LIST:
    case TepActionTypes.LOAD_TEP_ITEM_PRICE_DETAILS:

    case TepActionTypes.ADD_TEP_ITEM:
    case TepActionTypes.UPDATE_TEP_ITEM:
    case TepActionTypes.CANCEL:
    case TepActionTypes.CANCEL_TEP:
    case TepActionTypes.CONFIRM_TEP:
    case TepActionTypes.CONFIRM_TEP_REQUEST:
    case TepActionTypes.DELETE_TEP_ITEM:
    case TepActionTypes.LOAD_RSO_LIST:
    case TepActionTypes.LOAD_STUDDED_PRODUCT_DETAILS:
    case TepActionTypes.LOAD_TEP_ITEM_CODE_DETAILS:
    case TepActionTypes.LOAD_TEP_ITEM_DETAILS:
    case TepActionTypes.LOAD_TEP_TRANSACTION_DETAILS:
    case TepActionTypes.DELETE_TEP_TRANSACTION_DETAILS:
    case TepActionTypes.LOAD_WORKFLOW_DETAILS:
    case TepActionTypes.LOAD_REFUND_CASH_LIMIT:
    //case TepActionTypes.LOAD_TEP_ITEM_CUT_PIECE_DETAILS:
    case TepActionTypes.LOAD_CM_LIST_ITEM_TEP_CONFIGURATION:
    case TepActionTypes.LOAD_GOLD_PLUS_LOCATIONS:
    case TepActionTypes.LOAD_FTEP_REASONS:
    case TepActionTypes.LOAD_AVAILABLE_DISCOUNTS_LIST:
      return { ...state, isLoading: true, errors: null };
    case TepActionTypes.FILE_UPLOAD:
      return {
        ...state,
        isLoading: true,
        errors: null,
        uploadFileResponse: false
      };

    case TepActionTypes.UPDATE_TEP_ITEM_PRICE_DETAILS:
      return { ...state, isLoadingPriceUpdate: true, errors: null };

    case TepActionTypes.FILE_ID_PROOF_DOWNLOAD_URL:
      return {
        ...state,
        isLoading: true,
        errors: null,
        downloadIdProofFileUrl: null
      };

    case TepActionTypes.FILE_CANCELLED_CHEQUE_DOWNLOAD_URL:
      return {
        ...state,
        isLoading: true,
        errors: null,
        downloadCancelledChequeFileUrl: null
      };

    case TepActionTypes.FILE_APPROVAL_MAIL_DOWNLOAD_URL:
      return {
        ...state,
        isLoading: true,
        errors: null,
        downloadApprovalMailFileUrl: null
      };

    case TepActionTypes.FILE_UPLOAD_SUCCESS:
      return {
        ...state,
        uploadFileResponse: action.payload,
        errors: null,
        isLoading: false
      };

    case TepActionTypes.FILE_ID_PROOF_DOWNLOAD_URL_SUCCESS:
      return {
        ...state,
        downloadIdProofFileUrl: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.FILE_CANCELLED_CHEQUE_DOWNLOAD_URL_SUCCESS:
      return {
        ...state,
        downloadCancelledChequeFileUrl: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.FILE_APPROVAL_MAIL_DOWNLOAD_URL_SUCCESS:
      return {
        ...state,
        downloadApprovalMailFileUrl: action.payload,
        errors: null,
        isLoading: false
      };

    case TepActionTypes.FILE_UPLOAD_FAILURE:
      return {
        ...state,
        uploadFileResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.FILE_ID_PROOF_DOWNLOAD_URL_FAILURE:
      return {
        ...state,
        downloadIdProofFileUrl: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.FILE_CANCELLED_CHEQUE_DOWNLOAD_URL_FAILURE:
      return {
        ...state,
        downloadCancelledChequeFileUrl: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.FILE_APPROVAL_MAIL_DOWNLOAD_URL_FAILURE:
      return {
        ...state,
        downloadApprovalMailFileUrl: null,
        errors: action.payload,
        isLoading: false
      };

    case TepActionTypes.CREATE_OPEN_TEP_TRANSACTION_SUCCESS:
      return {
        ...state,
        createOpenTepTransactionResponse: action.payload,
        errors: null,
        isLoading: false,
        isOpenTaskLoading: false
      };
    case TepActionTypes.CREATE_OPEN_TEP_TRANSACTION_FAILURE:
      return {
        ...state,
        createOpenTepTransactionResponse: null,
        errors: action.payload,
        isLoading: false,
        isOpenTaskLoading: false
      };
    case TepActionTypes.UPDATE_OPEN_TEP_TRANSACTION_SUCCESS:
      return {
        ...state,
        updateOpenTepTransactionResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.UPDATE_OPEN_TEP_TRANSACTION_FAILURE:
      return {
        ...state,
        createOpenTepTransactionResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.GET_TEP_ITEM_CONFIGURATION_SUCCESS:
      return {
        ...state,
        tepItemConfiguratonResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.GET_TEP_ITEM_CONFIGURATION_FAILURE:
      return {
        ...state,
        tepItemConfiguratonResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.GET_TEP_ITEM_EXCEPTION_CONFIGURATION_SUCCESS:
      return {
        ...state,
        isExceptionScenario:
          action.payload?.tepOfferDetails?.offerDetails?.type ===
          'TEP_EXCEPTION_CONFIG',
        errors: null,
        isLoading: false
      };
    case TepActionTypes.GET_TEP_ITEM_EXCEPTION_CONFIGURATION_FAILURE:
      return {
        ...state,
        tepItemConfiguratonResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.GET_TEP_CASH_MEMO_ITEM_LIST_SUCCESS:
      return {
        ...state,
        tepCashMemoResponseItemList: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.GET_TEP_CASH_MEMO_ITEM_LIST_FAILURE:
      return {
        ...state,
        tepCashMemoResponseItemList: null,
        errors: action.payload,
        isLoading: false
      };

    case TepActionTypes.LOAD_WORKFLOW_DETAILS_FAILURE:
      return {
        ...state,
        errors: action.payload,
        selectedData: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_TEP_ITEM_PRICE_DETAILS_SUCCESS:
      return {
        ...state,
        tepPriceDetailsResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_TEP_ITEM_PRICE_DETAILS_FAILURE:
      return {
        ...state,
        tepPriceDetailsResponse: null,
        errors: action.payload,
        isLoading: false
      };

    case TepActionTypes.UPDATE_TEP_ITEM_PRICE_DETAILS_SUCCESS:
      return {
        ...state,
        UpdatePriceDetailsResponse: action.payload,
        errors: null,
        isLoadingPriceUpdate: false
      };
    case TepActionTypes.UPDATE_TEP_ITEM_PRICE_DETAILS_FAILURE:
      return {
        ...state,
        errors: action.payload,
        UpdatePriceDetailsResponse: null,
        isLoadingPriceUpdate: false
      };
    case TepActionTypes.ADD_TEP_ITEM_SUCCESS:
      return {
        ...state,
        addTepItemResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.ADD_TEP_ITEM_FAILURE:
      return {
        ...state,
        addTepItemResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.LOAD_WORKFLOW_DETAILS_SUCCESS:
      return {
        ...state,
        selectedData: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.UPDATE_TEP_ITEM_SUCCESS:
      return {
        ...state,
        addTepItemResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.UPDATE_TEP_ITEM_FAILURE:
      return {
        ...state,
        updateTepItemResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.CONFIRM_TEP_SUCCESS:
    case TepActionTypes.CONFIRM_TEP_REQUEST_SUCCESS:
      return {
        ...state,
        confirmTepItemResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.CONFIRM_TEP_FAILURE:
    case TepActionTypes.CONFIRM_TEP_REQUEST_FAILURE:
      return {
        ...state,
        confirmTepItemResponse: null,
        errors: action.payload,
        isLoading: false
      };

    case TepActionTypes.CANCEL_SUCCESS:
      return {
        ...state,
        isLoading: false,
        cancelResponse: action.payload,
        errors: null
      };

    case TepActionTypes.CANCEL_TEP_SUCCESS:
      return {
        ...state,
        isLoading: false,
        cancelTEPResponse: action.payload,
        errors: null
      };

    case TepActionTypes.CANCEL_FAILURE:
    case TepActionTypes.CANCEL_TEP_FAILURE:
      return {
        ...state,
        isLoading: false,
        errors: action.payload
      };

    case TepActionTypes.LOAD_REFUND_CASH_LIMIT_SUCCESS:
      return {
        ...state,
        isLoading: false,
        refundCashLimit: action.payload,
        errors: null
      };

    case TepActionTypes.LOAD_REFUND_CASH_LIMIT_FAILURE:
      return {
        ...state,
        isLoading: false,
        errors: action.payload
      };

    case TepActionTypes.DELETE_TEP_ITEM_SUCCESS:
      return {
        ...state,
        deleteTepItemResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.DELETE_TEP_ITEM_FAILURE:
      return {
        ...state,
        deleteTepItemResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.LOAD_RSO_LIST_SUCCESS:
      return {
        ...state,
        rsoList: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_RSO_LIST_FAILURE:
      return {
        ...state,
        rsoList: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.LOAD_STUDDED_PRODUCT_DETAILS_SUCCESS:
      return {
        ...state,
        studdedProductGroupCodes: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_STUDDED_PRODUCT_DETAILS_FAILURE:
      return {
        ...state,
        studdedProductGroupCodes: [],
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.LOAD_TEP_ITEM_CODE_DETAILS_SUCCESS:
      return {
        ...state,
        scannedTepItemCode: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_TEP_ITEM_CODE_DETAILS_FAILURE:
      return {
        ...state,
        scannedTepItemCode: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.LOAD_TEP_ITEM_DETAILS_SUCCESS:
      return {
        ...state,
        viewTepItemResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_TEP_ITEM_DETAILS_FAILURE:
      return {
        ...state,
        viewTepItemResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.LOAD_TEP_TRANSACTION_DETAILS_SUCCESS:
      return {
        ...state,
        viewTepTransactionResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_TEP_TRANSACTION_DETAILS_FAILURE:
      return {
        ...state,
        viewTepTransactionResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.DELETE_TEP_TRANSACTION_DETAILS_SUCCESS:
      return {
        ...state,
        deleteTepTransactionResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.DELETE_TEP_TRANSACTION_DETAILS_FAILURE:
      return {
        ...state,
        deleteTepTransactionResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.LOAD_CM_LIST_ITEM_TEP_CONFIGURATION_SUCCESS:
      return {
        ...state,
        cmListItemTepConfigurationResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_CM_LIST_ITEM_TEP_CONFIGURATION_FAILURE:
      return {
        ...state,
        cmListItemTepConfigurationResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.LOAD_GOLD_PLUS_LOCATIONS_SUCCESS:
      return {
        ...state,
        goldPlusLocations: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_GOLD_PLUS_LOCATIONS_FAILURE:
      return {
        ...state,
        goldPlusLocations: [],
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.LOAD_FTEP_REASONS_SUCCESS:
      return {
        ...state,
        fvtReasons: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_FTEP_REASONS_FAILURE:
      return {
        ...state,
        fvtReasons: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.UPDATE_TEP_TRANSACTION_PRICE_DETAILS:
      return {
        ...state,
        errors: null,
        isLoading: true
      };
    case TepActionTypes.UPDATE_TEP_TRANSACTION_PRICE_DETAILS_SUCCESS:
      return {
        ...state,
        updateTepTransactionPriceDetailsResponse: action.payload,
        errors: null,
        isLoading: false
      };
    case TepActionTypes.UPDATE_TEP_TRANSACTION_PRICE_DETAILS_FAILURE:
      return {
        ...state,
        updateTepTransactionPriceDetailsResponse: null,
        errors: action.payload,
        isLoading: false
      };
    case TepActionTypes.CREATE_OPEN_CUT_PIECE_TEP_TRANSACTION:
      return {
        ...state,
        errors: null,
        isLoading: true
      };
    case TepActionTypes.CREATE_OPEN_CUT_PIECE_TEP_TRANSACTION_SUCCESS:
      return {
        ...state,
        errors: null,
        createOpenCutPieceTepTransactionResponse: action.payload,
        isLoading: false
      };
    case TepActionTypes.CREATE_OPEN_CUT_PIECE_TEP_TRANSACTION_FAILURE:
      return {
        ...state,
        errors: action.payload,
        createOpenCutPieceTepTransactionResponse: null,
        isLoading: false
      };
    case TepActionTypes.PATCH_CUT_PIECE_TEP_TRANSACTION:
      return {
        ...state,
        errors: null,
        isLoading: true
      };
    case TepActionTypes.PATCH_CUT_PIECE_TEP_TRANSACTION_SUCCESS:
      return {
        ...state,
        errors: null,
        patchCutPieceTepTransactionResponse: action.payload,
        isLoading: false
      };
    case TepActionTypes.PATCH_CUT_PIECE_TEP_TRANSACTION_FAILURE:
      return {
        ...state,
        errors: action.payload,
        patchCutPieceTepTransactionResponse: null,
        isLoading: false
      };
    case TepActionTypes.ADD_CUT_PIECE_TEP_ITEM:
      return {
        ...state,
        errors: null,
        isLoading: true
      };
    case TepActionTypes.ADD_CUT_PIECE_TEP_ITEM_SUCCESS:
      return {
        ...state,
        errors: null,
        addCutPieceTepItemResponse: action.payload,
        isLoading: false
      };
    case TepActionTypes.ADD_CUT_PIECE_TEP_ITEM_FAILURE:
      return {
        ...state,
        errors: action.payload,
        addCutPieceTepItemResponse: null,
        isLoading: false
      };
    case TepActionTypes.PATCH_CUT_PIECE_TEP_ITEM:
      return {
        ...state,
        errors: null,
        isLoading: true
      };
    case TepActionTypes.PATCH_CUT_PIECE_TEP_ITEM_SUCCESS:
      return {
        ...state,
        errors: null,
        patchCutPieceTepItemResponse: action.payload,
        isLoading: false
      };
    case TepActionTypes.PATCH_CUT_PIECE_TEP_ITEM_FAILURE:
      return {
        ...state,
        errors: action.payload,
        patchCutPieceTepItemResponse: null,
        isLoading: false
      };
    case TepActionTypes.CONFIRM_CUT_PIECE_TEP_TRANSACTION:
      return {
        ...state,
        errors: null,
        isLoading: true
      };
    case TepActionTypes.CONFIRM_CUT_PIECE_TEP_TRANSACTION_SUCCESS:
      return {
        ...state,
        errors: null,
        confirmCutPieceTepItemResponse: action.payload,
        isLoading: false
      };
    case TepActionTypes.CONFIRM_CUT_PIECE_TEP_TRANSACTION_FAILURE:
      return {
        ...state,
        errors: action.payload,
        confirmCutPieceTepItemResponse: null,
        isLoading: false
      };
    case TepActionTypes.LOAD_AVAILABLE_DISCOUNTS_LIST_SUCCESS:
      return {
        ...state,
        errors: null,
        availableDiscountsList: action.payload,
        isLoading: false
      };
    case TepActionTypes.LOAD_AVAILABLE_DISCOUNTS_LIST_FAILURE:
      return {
        ...state,
        errors: action.payload,
        availableDiscountsList: null,
        isLoading: false
      };
    case TepActionTypes.SET_REMARKS:
      return {
        ...state,
        remarks: action.payload
      };
    case TepActionTypes.SET_HOLD_TRANSACCTION_METAL_RATES:
      return {
        ...state,
        holdTransactionMetalRates: action.payload
      };
    case TepActionTypes.SET_SELECTED_RSO_NAME:
      return {
        ...state,
        selectedRsoName: action.payload
      };
    case TepActionTypes.SET_TOTALQTY:
      return {
        ...state,
        totalQty: action.payload
      };
    case TepActionTypes.SET_TOTAL_GROSS_WT:
      return {
        ...state,
        totalGrossWt: action.payload
      };
    case TepActionTypes.SET_TOTAL_EXCHANGE_AMT:
      return {
        ...state,
        totalExchangeAmt: action.payload
      };
    case TepActionTypes.SELECTED_PAYMENT_METHOD:
      return {
        ...state,
        selectedPaymentMethod: action.payload
      };
    case TepActionTypes.SELECTED_TEP_TYPE:
      return {
        ...state,
        selectedTepType: action.payload
      };
    case TepActionTypes.SET_CUT_PIECE_TOTAL_QTY:
      return {
        ...state,
        cutPieceTotalQty: action.payload
      };
    case TepActionTypes.SET_CUT_PIECE_TOTAL_VALUE:
      return {
        ...state,
        cutPieceTotalValue: action.payload
      };
    case TepActionTypes.SET_IS_REFUND_FORM_VALID:
      return {
        ...state,
        isRefundFormValid: action.isValid
      };
    case TepActionTypes.SET_IS_REQUEST_RAISING_SCENARIO:
      return {
        ...state,
        isRequestRaisingScenario: action.payload
      };
    case TepActionTypes.RESET_TEP:
      return {
        errors: null,
        isLoading: false,
        isOpenTaskLoading: false,
        selectedRsoName: null,
        selectedCmItem: null,
        createOpenTepTransactionResponse: null,
        updateOpenTepTransactionResponse: null,
        tepItemConfiguratonResponse: null,
        tepCashMemoResponseItemList: null,
        cancelResponse: null,
        tepPriceDetailsResponse: null,
        addTepItemResponse: null,
        UpdatePriceDetailsResponse: null,
        updateTepItemResponse: null,
        confirmTepItemResponse: null,
        deleteTepItemResponse: null,
        rsoList: [],
        remarks: '',
        totalQty: 0,
        totalGrossWt: 0,
        totalExchangeAmt: 0,
        selectedPaymentMethod: null,
        selectedTepType: null,
        scannedTepItemCode: null,
        selectedData: null,
        viewTepTransactionResponse: null,
        viewTepItemResponse: null,
        deleteTepTransactionResponse: null,
        tepItemCutPieceDetailsResponse: null,
        cutPieceTotalQty: 0,
        cutPieceTotalValue: 0,
        cmListItemTepConfigurationResponse: null,
        isRefundFormValid: false,
        cancelTEPResponse: null,
        isRequestRaisingScenario: false,
        goldPlusLocations: [],
        uploadFileResponse: false,
        downloadIdProofFileUrl: null,
        downloadCancelledChequeFileUrl: null,
        downloadApprovalMailFileUrl: null,
        fvtReasons: [],
        updateTepTransactionPriceDetailsResponse: null,
        isLoadingPriceUpdate: false,
        createOpenCutPieceTepTransactionResponse: null,
        patchCutPieceTepTransactionResponse: null,
        addCutPieceTepItemResponse: null,
        patchCutPieceTepItemResponse: null,
        confirmCutPieceTepItemResponse: null,
        availableDiscountsList: null,
        holdTransactionMetalRates: null,
        refundCashLimit: null
      };
    default:
      return state;
  }
}
